Add function to extract sources and add unit testing
