+++ tests/fixtures/testrepo/upstream/packages/hello/hello.spec
@@ -18,6 +18,8 @@ BuildRequires: gettext
 Requires(post): info
 Requires(preun): info
 
+Obsoletes: goodbye
+
 %description
 The "Hello World" program, done with all bells and whistles of a proper FOSS
 project, including configuration, build, internationalization, help files, etc.
